+++ lib/sound_util/sink.rb
@@ -3,5 +3,6 @@
 module SoundUtil
   module Sink
     autoload :Playback, "sound_util/sink/playback"
+    autoload :Preview, "sound_util/sink/preview"
   end
 end